+++ src/ezdxf/addons/drawing/svg.py
@@ -6,7 +6,7 @@ from typing import Iterable, Sequence, no_type_check
 import copy
 from xml.etree import ElementTree as ET
 
-from ezdxf.math import Vec2
+from ezdxf.math import Vec2, BoundingBox2d
 from ezdxf.path import Command
 
 
@@ -21,7 +21,8 @@ __all__ = ["SVGBackend"]
 
 class SVGBackend(recorder.Recorder):
     """This is a native SVG rendering backend and does not require any external packages
-    to render SVG images other than the core dependencies.
+    to render SVG images other than the core dependencies.  This backend support content
+    cropping at page margins.
     """
 
     def __init__(self) -> None:
@@ -29,24 +30,43 @@ class SVGBackend(recorder.Recorder):
         self._init_flip_y = True
 
     def get_xml_root_element(
-        self, page: layout.Page, settings: layout.Settings = layout.Settings()
+        self,
+        page: layout.Page,
+        *,
+        settings: layout.Settings = layout.Settings(),
+        render_box: BoundingBox2d | None = None,
     ) -> ET.Element:
+        top_origin = True
         settings = copy.copy(settings)
-        # The DXF coordinates are mapped to integer viewBox coordinates in the first
+        # DXF coordinates are mapped to integer viewBox coordinates in the first
         # quadrant, producing compact SVG files. The larger the coordinate range, the
         # more precise and the lager the files.
         settings.output_coordinate_space = 1_000_000
 
         # This player changes the original recordings!
         player = self.player()
+        if render_box is None:
+            render_box = player.bbox()
 
-        output_layout = layout.Layout(player.bbox(), flip_y=self._init_flip_y)
+        # the page origin (0, 0) is in the top-left corner.
+        output_layout = layout.Layout(render_box, flip_y=self._init_flip_y)
         page = output_layout.get_final_page(page, settings)
         if page.width == 0 or page.height == 0:
             return ET.Element("svg")  # empty page
 
-        m = output_layout.get_placement_matrix(page, settings)
+        m = output_layout.get_placement_matrix(
+            page, settings=settings, top_origin=top_origin
+        )
+        # transform content to the output coordinates space:
         player.transform(m)
+        if settings.crop_at_margins:
+            p1, p2 = page.get_margin_rect(top_origin=top_origin)  # in mm
+            # scale factor to map page coordinates to output space coordinates:
+            output_scale = settings.page_output_scale_factor(page)
+            max_sagitta = 0.1 * output_scale  # curve approximation 0.1 mm
+            # crop content inplace by the margin rect:
+            player.crop_rect(p1 * output_scale, p2 * output_scale, max_sagitta)
+
         self._init_flip_y = False
         backend = self.make_backend(page, settings)
         player.replay(backend)
@@ -55,7 +75,9 @@ class SVGBackend(recorder.Recorder):
     def get_string(
         self,
         page: layout.Page,
+        *,
         settings: layout.Settings = layout.Settings(),
+        render_box: BoundingBox2d | None = None,
         xml_declaration=True,
     ) -> str:
         """Returns the XML data as unicode string.
@@ -63,11 +85,14 @@ class SVGBackend(recorder.Recorder):
         Args:
             page: page definition, see :class:`~ezdxf.addons.drawing.layout.Page`
             settings: layout settings, see :class:`~ezdxf.addons.drawing.layout.Settings`
+            render_box: set explicit region to render, default is content bounding box
             xml_declaration: inserts the "<?xml version='1.0' encoding='utf-8'?>" string
                 in front of the <svg> element
 
         """
-        xml = self.get_xml_root_element(page, settings)
+        xml = self.get_xml_root_element(
+            page, settings=settings, render_box=render_box
+        )
         return ET.tostring(xml, encoding="unicode", xml_declaration=xml_declaration)
 
     @staticmethod