+++ src/ezdxf/entities/attrib.py
@@ -2,9 +2,10 @@
 # License: MIT License
 from __future__ import annotations
 from typing import TYPE_CHECKING, Optional
+from typing_extensions import Self
 import copy
 from ezdxf.lldxf import validator
-from ezdxf.math import NULLVEC, Vec3, Z_AXIS, OCS
+from ezdxf.math import NULLVEC, Vec3, Z_AXIS, OCS, Matrix44
 from ezdxf.lldxf.attributes import (
     DXFAttr,
     DXFAttributes,
@@ -381,6 +382,16 @@ class BaseAttrib(Text):
         if self._embedded_mtext and clone._embedded_mtext:
             self._embedded_mtext.map_resources(clone._embedded_mtext, mapping)
         # todo: map handles in embedded XRECORD if a real world example shows up
+
+    def transform(self, m: Matrix44) -> Self:
+        if self._embedded_mtext is None:
+            super().transform(m)
+        else:
+            mtext = self._embedded_mtext.virtual_mtext_entity()
+            mtext.transform(m)
+            self.set_mtext(mtext, graphic_properties=False)
+            self.post_transform(m)
+        return self
 
 
 def _update_content_from_mtext(text: Text, mtext: MText) -> None: